Cambios a los nombres de archivos de la carpeta controllers y cambio a la propiedad del dto de libro nove

+++ NeoAPI/DTOs/LibroNovedades/LibroNoveDTO.cs
@@ -14,7 +14,7 @@ public class LibroNoveDTO
 
     public string Lndiscrepa { get; set; } = null!;
 
-    public int LntiePerMi { get; set; }
+    public double LntiePerMi { get; set; }
 
     public string LnfichaRes { get; set; } = null!;
